[refactor] move us trend query file to queries folder under utilities
